spotify-downloader/requirements.txt
spotipy
yt-dlp
python-dotenv

